+++ mobile/lib/views/widgets/symptoms/custom_symptom_widget.dart
@@ -11,7 +11,7 @@ import 'package:get/get.dart';
 class CustomSymptomWidget extends StatelessWidget {
   final int symptomID;
   final String label;
-  final int value;
+  final double value;
   final Function onUpdate;
 
   CustomSymptomWidget({
@@ -43,7 +43,7 @@ class CustomSymptomWidget extends StatelessWidget {
 
     // Добавляем слушатель для обновления реактивного значения в контроллере
     valueInputController.addListener(() {
-      int? newValue = int.tryParse(valueInputController.text);
+      double? newValue =  double.tryParse(valueInputController.text);
       if (newValue != null) {
         controller.updateSymptomValueInDB(symptomID, newValue);
       }
